(ticks) Peel apart CPU emulation from main()
Remove .sna loading etc - may come back later on of course...
