+++ src/ticks/Makefile
@@ -5,7 +5,7 @@ OBJS = ticks.o rxk_inst.o r4k_inst.o i8085_inst.o z80n_inst.o gbz80_inst.o \
 	hook_cpm.o hook_console.o hook_io.o hook_misc.o hook.o debugger.o breakpoints.o \
 	profiler.o exp_engine.o debugger_ticks.o linenoise.o utf8.o syms.o \
 	disassembler_alg.o memory.o am9511.o acia.o hook_rc2014.o debug.o srcfile.o \
-	i8255.o \
+	i8255.o ticks_main.o \
 	../common/dirname.o $(UNIXem_OBJS)
 GDBOBJS = cpu.o backend.o syms.o disassembler_alg.o debug.o exp_engine.o debugger.o breakpoints.o profiler.o debugger_gdb.o debugger_mi2.o debugger_gdb_packets.o linenoise.o srcfile.o ../common/dirname.o sxmlc.o sxmlsearch.o $(UNIXem_OBJS)
 DISOBJS = disassembler_main.o syms.o disassembler_alg.o debug.o exp_engine.o backend.o
