+++ win32/ticks/ticks.vcxproj
@@ -201,6 +201,7 @@
     <ClCompile Include="..\..\src\ticks\srcfile.c" />
     <ClCompile Include="..\..\src\ticks\syms.c" />
     <ClCompile Include="..\..\src\ticks\ticks.c" />
+    <ClCompile Include="..\..\src\ticks\ticks_main.c" />
     <ClCompile Include="..\..\src\ticks\utf8.c" />
     <ClCompile Include="..\..\src\ticks\z180_inst.c" />
     <ClCompile Include="..\..\src\ticks\z80n_inst.c" />
